fix: move build to build-js
