+++ webpack.config.js
@@ -3,7 +3,9 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 const webpackConfig = require('@nextcloud/webpack-vue-config')
-const WebpackSPDXPlugin = require('./build/WebpackSPDXPlugin.js')
+// eslint-disable-next-line n/no-extraneous-require
+const TerserPlugin = require('terser-webpack-plugin')
+const WebpackSPDXPlugin = require('./build-js/WebpackSPDXPlugin.js')
 const path = require('path')
 
 webpackConfig.entry.workersrc = path.resolve(path.join('src', 'workersrc.js'))
@@ -11,6 +13,15 @@ webpackConfig.entry.public = path.resolve(path.join('src', 'public.js'))
 
 // keep pdfjs vendor in the js folder
 webpackConfig.output.clean = false
+
+webpackConfig.optimization.minimizer = [new TerserPlugin({
+	extractComments: false,
+	terserOptions: {
+		format: {
+			comments: false,
+		},
+	},
+})]
 
 webpackConfig.plugins = [
 	...webpackConfig.plugins,
